requirements.txt
streamlit
pdfplumber
python-docx
spacy
openai
keybert
yake
python-dotenv
